添加许多注释

+++ java/video/src/main/java/ljl/bilibili/video/service/video_production/impl/UploadAndEditServiceImpl.java
@@ -51,7 +51,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import static ljl.bilibili.video.constant.Constant.*;
-
+import static ljl.bilibili.video.constant.Constant.uploadPartMap;
+//上传视频
 @Slf4j
 @Service
 public class UploadAndEditServiceImpl implements UploadAndEditService {
@@ -70,24 +71,23 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
     AtomicBoolean uploadVideoSuccess = new AtomicBoolean(true);
     @Resource
     RedisTemplate objectRedisTemplate;
-    public static Map<String, UploadPart> uploadPartMap=new HashMap<>();
 
     /**
-     * 判断是否恶意文件、上传视频到minio、新增视频与视频数据记录、推送视频动态、发送数据同步消息
+     * 上传总片
      */
     @Override
     @Transactional
     public Result<Boolean> uploadTotal(UploadVideoRequest uploadVideoRequest) {
         try {
+            //将请求中的视频参数封装一部分
             String path = Files.createTempDirectory(".tmp").toString();
-            File file = new File(path, "test");
             Video video = uploadVideoRequest.toEntity();
             LambdaQueryWrapper<Video> queryWrapper = new LambdaQueryWrapper<>();
-            MultipartFile videoFile = null;
             String coverFile = uploadVideoRequest.getVideoCover();
             String url = "https://labilibili.com/video/" + uploadVideoRequest.getUrl();
             video.setUrl(url);
-            if (coverFile != null && uploadPartMap.get(uploadVideoRequest.getUserId()) == null) {
+            //如果视频封面不为空那么直接上传视频
+            if ((coverFile != null&&coverFile!="") && uploadPartMap.get(uploadVideoRequest.getUserId()) == null) {
 //                hasCover=true;
                 String prefixPath = "https://labilibili.com/video-cover/";
                 byte[] decodedBytes = java.util.Base64.getDecoder().decode(coverFile);
@@ -95,37 +95,32 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
                 String imgContentType = "image/jpeg";
                 video.setCover(prefixPath + coverFileName);
                 CustomMultipartFile coverMultipartFile = new CustomMultipartFile(decodedBytes, coverFileName, imgContentType);
-                CompletableFuture.runAsync(() -> log.info(""));
-                CompletableFuture.runAsync(() -> {
-                    log.info("");
-                    log.info("");
-                });
                 queryWrapper.eq(Video::getCover, UUID.randomUUID().toString().substring(0, 8) + coverFileName);
                 CompletableFuture<Void> uploadImgFuture = CompletableFuture.runAsync(() -> minioService.uploadImgFile(coverFileName, coverMultipartFile.getInputStream(), imgContentType)).handle((result, ex) -> {
+                    //异步回调通知异步任务是否执行成功
                     if (ex != null) {
                         uploadVideoSuccess.set(false);
-                        log.error("uploadimgfail");
+                        log.error("uploadImgFail");
                     }
                     return null;
                 });
-                log.info("封面" + video.getCover());
-
-
+                //发送转码通知并设置已有封面
                 CompletableFuture thenRunFuture = uploadImgFuture.thenRun(() ->
                         client.sendUploadNotice(new UploadVideo().setVideoId(video.getId()).setVideoName(video.getName()).setUrl(url).setHasCover(true)));
                 CompletableFuture<Void> sendNoticeFuture = CompletableFuture.runAsync(() -> {
                     User user = userMapper.selectById(uploadVideoRequest.getUserId());
+                    //发送生成关注up主动态视频的通知
                     client.dynamicNotice(uploadVideoRequest.toCoverDynamic(user, video));
-                    log.info("dynamicupload");
+                    log.info("dynamicUpload");
                 }).handle((result, ex) -> {
                     if (ex != null) {
-                        log.error("dynamicfail");
+                        log.error("dynamicFail");
                         uploadVideoSuccess.set(false);
                     }
                     return null;
                 });
             } else {
-
+                //发送生成关注up主动态视频的通知
                 CompletableFuture thenRunFuture = CompletableFuture.runAsync(() ->
                         client.sendUploadNotice(new UploadVideo().setVideoId(video.getId()).setVideoName(video.getName()).setUrl(url).setHasCover(false)));
                 CompletableFuture<Void> sendNoticeFuture = CompletableFuture.runAsync(() -> {
@@ -139,10 +134,9 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
                     return null;
                 });
             }
-            log.info(video.getCover());
             videoMapper.insert(video);
-            log.info(video.getCover());
             videoDataMapper.insert(new VideoData().setVideoId(video.getId()));
+            //发送MySQL与es数据同步的消息
             CompletableFuture<Void> sendDBChangeNotice = CompletableFuture.runAsync(() -> {
                 ObjectMapper objectMapper = new ObjectMapper();
                 JavaTimeModule module = new JavaTimeModule();
@@ -158,21 +152,12 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
                 map.remove(TABLE_ID);
                 client.sendDBChangeNotice(map);
             });
-
-//            CompletableFuture.allOf(uploadVideoFuture, sendNoticeFuture).join();
-//            if (!uploadVideoSuccess.get()) {
-//                log.error("lose");
-//            } else {
-//                // 所有任务成功
-//                log.info("ok");
-//            }
             return Result.success(true);
         } catch (Exception e) {
             e.printStackTrace();
             return Result.error("寄");
         }
     }
-
     /**
      * 编辑视频并发送数据同步消息
      */
@@ -190,6 +175,7 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
             map.put(TABLE_NAME, VIDEO_TABLE_NAME);
             map.put(OPERATION_TYPE, OPERATION_TYPE_UPDATE);
             Video video = editVideoRequest.toEntity();
+            //如果视频文件不为空，则该视频连视频文件都修改了
             if (videoFile != null) {
                 String videoUrl = UUID.randomUUID().toString().substring(0, 10) + editVideoRequest.getName();
                 video.setUrl(videoUrl);
@@ -202,6 +188,7 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
                     }
 
                 });
+                //异步发送转码消息时查询redis进行限流，如果已有三个正在进行的转码则不不发送转码消息等待转码完毕再发送
                 CompletableFuture<Void> future1 = CompletableFuture.runAsync(() -> {
                     while (true) {
                         String key = "encode-count";
@@ -223,6 +210,7 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
                 });
             }
             MultipartFile coverFile = editVideoRequest.getCover();
+            //如果视频文件为空则只需要修改视频相关信息
             if (coverFile != null) {
                 CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(() -> {
                     String coverUrl = UUID.randomUUID().toString().substring(0, 10) + coverFile.getOriginalFilename();
@@ -274,50 +262,39 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
     }
 
     /**
-     * 上传视频时获取视频封面
+     * 上传视频的分片、合并分片并获取最后截取的封面和存储到minio中的路径方便上传视频相关信息时使用
      */
     @Override
     public Result<List<String>> uploadPart(UploadPartRequest uploadPartRequest) throws IOException, EncoderException, ServerException, InsufficientDataException, ErrorResponseException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
-
+        //前端使用框架会在上传分片时将某个传参属性值重复两遍赋值到该属性中，因此需要去掉后续的传参以防映射不对
+        int commaIndex = uploadPartRequest.getResumableIdentifier().indexOf(',');
+        uploadPartRequest.setResumableIdentifier(uploadPartRequest.getResumableIdentifier().substring(0,commaIndex));
         String resumableIdentifier = uploadPartRequest.getResumableIdentifier();
         String videoName = "";
         String cover = "";
-        if(uploadPartRequest.getResumableChunkNumber()==1){
-            String path = Files.createTempDirectory(".tmp").toString();
-            File file = new File(path, "test");
-            InputStream videoFileInputStream = uploadPartRequest.getFile().getInputStream();
-            byte[] bytes=IoUtil.readBytes(videoFileInputStream);
-            ByteArrayInputStream byteArrayInputStream=new ByteArrayInputStream(bytes);
-            Files.copy(byteArrayInputStream, Paths.get(file.getAbsolutePath()), StandardCopyOption.REPLACE_EXISTING);
-            if (!FileTypeUtil.getType(file).equals("mp4")) {
-                file.delete();
-                return Result.error("上传恶意文件");
-            } else {
-                file.delete();
-                log.info("文件无问题");
-            }
-        }
-        if(uploadPartMap.get(resumableIdentifier).getTotalCount().equals(uploadPartRequest.getResumableTotalChunks()-1)){
+        //当视频上传的分片到最后一片时截取该分片的图片出来并将值传到cover里
+        if(uploadPartMap.get(resumableIdentifier)!=null&&uploadPartMap.get(resumableIdentifier).getTotalCount().equals(uploadPartRequest.getResumableTotalChunks()-1)){
             InputStream videoFileInputStream = uploadPartRequest.getFile().getInputStream();
             byte[] bytes=IoUtil.readBytes(videoFileInputStream);
             ByteArrayInputStream byteArrayInputStream=new ByteArrayInputStream(bytes);
             String filePath = Files.createTempDirectory(".tmp").toString();
-            String coverFileName="coverFileName.jpg";
+            String coverFileName=UUID.randomUUID().toString().substring(0,10)+".jpg";
             String videoFileName="video";
             File videoFile=new File(filePath,videoFileName);
-            File coverFile1 = new File(filePath, coverFileName);
+            File coverFile = new File(filePath, coverFileName);
             Files.copy(byteArrayInputStream, Paths.get(videoFile.getAbsolutePath()), StandardCopyOption.REPLACE_EXISTING);
             ScreenExtractor screenExtractor = new ScreenExtractor();
             MultimediaObject multimediaObject = new MultimediaObject(videoFile);
-            screenExtractor.renderOneImage(multimediaObject, -1, -1, 1000, coverFile1, 1);
-            InputStream inputStream1 = new FileInputStream(coverFile1);
-            cover=Base64.encode(IoUtil.readBytes(inputStream1));
+            screenExtractor.renderOneImage(multimediaObject, -1, -1, 1000, coverFile, 1);
+            InputStream inputStream = new FileInputStream(coverFile);
+            cover=Base64.encode(IoUtil.readBytes(inputStream));
             videoFile.delete();
-            coverFile1.delete();
+            coverFile.delete();
         }
+        //minio中同名文件是会覆盖的，因此若出现两个人上传同名文件后一个人上传的会把前一个上传的覆盖，所以需要给视频加独特后缀区分开
         String name = resumableIdentifier + UUID.randomUUID().toString().substring(0, 10);
         minioService.uploadVideoFile(name, uploadPartRequest.getFile().getInputStream(), VIDEO_TYPE);
-        log.info(uploadPartMap.toString());
+        //将用户之前上传的分片信息缓存起来，方便后续断点续传时发送请求确认是否已上传过分片，确认了上传过的片就不会再上传
         Map<Integer, String> newUploadPartMap = uploadPartMap.getOrDefault(resumableIdentifier, new UploadPart()).getPartMap();
         newUploadPartMap.put(uploadPartRequest.getResumableChunkNumber(), name);
         UploadPart uploadPart=uploadPartMap.getOrDefault(resumableIdentifier,new UploadPart());
@@ -328,7 +305,6 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
             log.info("合并");
             videoName = resumableIdentifier + UUID.randomUUID().toString().substring(0, 10);
             minioService.composePart(resumableIdentifier, videoName);
-            cover="https://labilibili.com/video/山海2549ec9d-c";
         }
         List<String> list = new ArrayList<>();
         list.add(videoName);
@@ -337,15 +313,12 @@ public class UploadAndEditServiceImpl implements UploadAndEditService {
     }
     @Override
     public ResponseEntity<Result<Boolean>> getProcessor(String resumableIdentifier, Integer resumableChunkNumber) {
-        List<Integer> indexList = new ArrayList<>();
-        if(uploadPartMap.getOrDefault(resumableIdentifier,new UploadPart()).getPartMap().size()>0){
-            for (Map.Entry<Integer, String> entry : uploadPartMap.get(resumableIdentifier).getPartMap().entrySet()) {
-                indexList.add(entry.getKey());
+        //根据之前存的映射判断，如果上传过了片就不用再上传了
+        for (Map.Entry<Integer, String> entry : uploadPartMap.getOrDefault(resumableIdentifier,new UploadPart()).getPartMap().entrySet()) {
+            if(entry.getKey().intValue()==(resumableChunkNumber.intValue())){
+                return ResponseEntity.ok(Result.data(true));
             }
         }
-        if(!indexList.contains(resumableChunkNumber)){
-            return ResponseEntity.noContent().build();
-        }
-        return ResponseEntity.ok(Result.data(true));
+        return ResponseEntity.noContent().build();
     }
 }
